more modal refacto
